bugfix(#545): Block distributions cause segfault
- Due to some copy pasta, passing double to a param that takes int, etc. type
  errors.

- Also updated to address new libra compiler warning definitions and re-ran
  formatter, including adding the "final" specifier whenever possible to enable
  divirtualization.

+semver: bugfix

+++ include/fordyca/ds/arena_grid.hpp
@@ -50,8 +50,8 @@ class arena_grid : public rcppsw::ds::stacked_grid<arena_layer_stack> {
   using view = rcppsw::ds::base_grid2D<ds::cell2D>::grid_view;
   using const_view = rcppsw::ds::base_grid2D<ds::cell2D>::const_grid_view;
 
-  constexpr static uint kCell = 0;
-  constexpr static uint kRobotOccupancy = 1;
+  constexpr static size_t kCell = 0;
+  constexpr static size_t kRobotOccupancy = 1;
 
   /**
    *
@@ -64,8 +64,8 @@ class arena_grid : public rcppsw::ds::stacked_grid<arena_layer_stack> {
    */
   arena_grid(double resolution, size_t x_max, size_t y_max)
       : stacked_grid(resolution, x_max, y_max) {
-    for (uint i = 0; i < xdsize(); ++i) {
-      for (uint j = 0; j < ydsize(); ++j) {
+    for (size_t i = 0; i < xdsize(); ++i) {
+      for (size_t j = 0; j < ydsize(); ++j) {
         access<kCell>(i, j).loc(rmath::vector2u(i, j));
       } /* for(j..) */
     }   /* for(i..) */
@@ -76,8 +76,8 @@ class arena_grid : public rcppsw::ds::stacked_grid<arena_layer_stack> {
     * blocks as well as setting all cells back to an empty state.
     */
   void reset(void) {
-    for (uint i = 0; i < xdsize(); ++i) {
-      for (uint j = 0; j < ydsize(); ++j) {
+    for (size_t i = 0; i < xdsize(); ++i) {
+      for (size_t j = 0; j < ydsize(); ++j) {
         access<kCell>(i, j).reset();
       } /* for(j..) */
     }   /* for(i..) */
